drm/openchrome: Hardware cursor fix for single HI devices
Older generation VIA Technologies Chrome IGP devices have only one
hardware icon (HI). Due to the way the code was written previously, it
was assuming the availability of two HIs, and as a result, if the user
was using older generation devices, no mouse cursor will be displayed
on the screen. This commit will add the necessary code to support older
devices with only one HI. The code was validated on ECS K8M890M-M
mainboard with K8M890 chipset.

Signed-off-by: Kevin Brace <kevinbrace@gmx.com>